Remove components namespace and folders
