+++ hardware_interface/include/hardware_interface/system_interface.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef HARDWARE_INTERFACE__COMPONENTS__ACTUATOR_INTERFACE_HPP_
-#define HARDWARE_INTERFACE__COMPONENTS__ACTUATOR_INTERFACE_HPP_
+#ifndef HARDWARE_INTERFACE__SYSTEM_INTERFACE_HPP_
+#define HARDWARE_INTERFACE__SYSTEM_INTERFACE_HPP_
 
 #include <memory>
 #include <string>
@@ -23,50 +23,51 @@
 #include "hardware_interface/hardware_info.hpp"
 #include "hardware_interface/types/hardware_interface_return_values.hpp"
 #include "hardware_interface/types/hardware_interface_status_values.hpp"
+#include "hardware_interface/visibility_control.h"
 
 namespace hardware_interface
 {
-namespace components
-{
 
 /**
-  * \brief Virtual Class to implement when integrating a 1 DoF actuator into ros2_control.
-  * The typical examples are conveyors or motors.
-  */
-class ActuatorInterface
+* \brief Virtual Class to implement when integrating a complex system into ros2_control.
+* The common examples for these types of hardware are multi-joint systems with or without sensors
+* such as industrial or humanoid robots.
+*/
+class SystemInterface
 {
 public:
-  ActuatorInterface() = default;
+  SystemInterface() = default;
 
   virtual
-  ~ActuatorInterface() = default;
+  ~SystemInterface() = default;
 
   /**
-   * \brief Configuration of the actuator from data parsed from the robot's URDF.
+   * \brief Configuration of the system from data parsed from the robot's URDF.
    *
-   * \param actuator_info structure with data from URDF.
+   * \param system_info structure with data from URDF.
    * \return return_type::OK if required data are provided and can be parsed,
    * return_type::ERROR otherwise.
    */
   virtual
-  return_type configure(const HardwareInfo & actuator_info) = 0;
+  return_type configure(const HardwareInfo & system_info) = 0;
 
-  /// Exports all state interfaces for this actuator.
+  /// Exports all state interfaces for this system.
   /**
    * The state interfaces have to be created and transfered according
-   * to the actuator info passed in for the configuration.
+   * to the system info passed in for the configuration.
    *
    * Note the ownership over the state interfaces is transfered to the caller.
+   *
    *
    * \return vector of state interfaces
    */
   virtual
   std::vector<StateInterface> export_state_interfaces() = 0;
 
-  /// Exports all command interfaces for this actuator.
+  /// Exports all command interfaces for this system.
   /**
    * The command interfaces have to be created and transfered according
-   * to the actuator info passed in for the configuration.
+   * to the system info passed in for the configuration.
    *
    * Note the ownership over the state interfaces is transfered to the caller.
    *
@@ -92,7 +93,7 @@ public:
   return_type stop() = 0;
 
   /**
-   * \brief Get name of the actuator hardware.
+   * \brief Get name of the system hardware.
    *
    * \return std::string name.
    */
@@ -100,14 +101,14 @@ public:
   std::string get_name() const = 0;
 
   /**
-   * \brief Get current state of the actuator hardware.
+   * \brief Get current state of the system hardware.
    *
    * \return status current status.
    */
   virtual
   status get_status() const = 0;
 
-  /// Read the current state values from the actuator.
+  /// Read the current state values from the actuators and sensors within the system.
   /**
    * The data readings from the physical hardware has to be updated
    * and reflected accordingly in the exported state interfaces.
@@ -118,7 +119,7 @@ public:
   virtual
   return_type read() = 0;
 
-  /// Write the current command values to the actuator.
+  /// Write the current command values to the actuator within the system.
   /**
    * The physical hardware shall be updated with the latest value from
    * the exported command interfaces.
@@ -129,6 +130,5 @@ public:
   return_type write() = 0;
 };
 
-}  // namespace components
 }  // namespace hardware_interface
-#endif  // HARDWARE_INTERFACE__COMPONENTS__ACTUATOR_INTERFACE_HPP_
+#endif  // HARDWARE_INTERFACE__SYSTEM_INTERFACE_HPP_
