+++ hardware_interface/src/resource_manager.cpp
@@ -20,12 +20,12 @@
 #include <unordered_map>
 #include <utility>
 
-#include "hardware_interface/components/actuator.hpp"
-#include "hardware_interface/components/actuator_interface.hpp"
-#include "hardware_interface/components/sensor.hpp"
-#include "hardware_interface/components/sensor_interface.hpp"
-#include "hardware_interface/components/system.hpp"
-#include "hardware_interface/components/system_interface.hpp"
+#include "hardware_interface/actuator.hpp"
+#include "hardware_interface/actuator_interface.hpp"
+#include "hardware_interface/sensor.hpp"
+#include "hardware_interface/sensor_interface.hpp"
+#include "hardware_interface/system.hpp"
+#include "hardware_interface/system_interface.hpp"
 #include "hardware_interface/component_parser.hpp"
 
 #include "pluginlib/class_loader.hpp"
@@ -38,11 +38,11 @@ class ResourceStorage
   static constexpr const char * pkg_name = "hardware_interface";
 
   static constexpr const char * actuator_interface_name =
-    "hardware_interface::components::ActuatorInterface";
+    "hardware_interface::ActuatorInterface";
   static constexpr const char * sensor_interface_name =
-    "hardware_interface::components::SensorInterface";
+    "hardware_interface::SensorInterface";
   static constexpr const char * system_interface_name =
-    "hardware_interface::components::SystemInterface";
+    "hardware_interface::SystemInterface";
 
 public:
   ResourceStorage()
@@ -98,7 +98,7 @@ public:
     const HardwareInfo & hardware_info,
     std::unordered_map<std::string, bool> & claimed_command_interface_map)
   {
-    initialize_hardware<components::Actuator, components::ActuatorInterface>(
+    initialize_hardware<Actuator, ActuatorInterface>(
       hardware_info, actuator_loader_, actuators_);
     if (return_type::OK != actuators_.back().configure(hardware_info)) {
       throw std::runtime_error(std::string("failed to configure ") + hardware_info.name);
@@ -109,7 +109,7 @@ public:
 
   void initialize_sensor(const HardwareInfo & hardware_info)
   {
-    initialize_hardware<components::Sensor, components::SensorInterface>(
+    initialize_hardware<Sensor, SensorInterface>(
       hardware_info, sensor_loader_, sensors_);
     if (return_type::OK != sensors_.back().configure(hardware_info)) {
       throw std::runtime_error(std::string("failed to configure ") + hardware_info.name);
@@ -121,7 +121,7 @@ public:
     const HardwareInfo & hardware_info,
     std::unordered_map<std::string, bool> & claimed_command_interface_map)
   {
-    initialize_hardware<components::System, components::SystemInterface>(
+    initialize_hardware<System, SystemInterface>(
       hardware_info, system_loader_, systems_);
     if (return_type::OK != systems_.back().configure(hardware_info)) {
       throw std::runtime_error(std::string("failed to configure ") + hardware_info.name);
@@ -131,13 +131,13 @@ public:
   }
 
   // hardware plugins
-  pluginlib::ClassLoader<components::ActuatorInterface> actuator_loader_;
-  pluginlib::ClassLoader<components::SensorInterface> sensor_loader_;
-  pluginlib::ClassLoader<components::SystemInterface> system_loader_;
+  pluginlib::ClassLoader<ActuatorInterface> actuator_loader_;
+  pluginlib::ClassLoader<SensorInterface> sensor_loader_;
+  pluginlib::ClassLoader<SystemInterface> system_loader_;
 
-  std::vector<components::Actuator> actuators_;
-  std::vector<components::Sensor> sensors_;
-  std::vector<components::System> systems_;
+  std::vector<Actuator> actuators_;
+  std::vector<Sensor> sensors_;
+  std::vector<System> systems_;
 
   std::unordered_map<std::string, status> hardware_status_map_;
 
@@ -257,10 +257,10 @@ bool ResourceManager::command_interface_exists(const std::string & key) const
          resource_storage_->command_interface_map_.end();
 }
 
-void ResourceManager::import_component(std::unique_ptr<components::ActuatorInterface> actuator)
+void ResourceManager::import_component(std::unique_ptr<ActuatorInterface> actuator)
 {
   resource_storage_->actuators_.emplace_back(
-    components::Actuator(std::move(actuator)));
+    Actuator(std::move(actuator)));
   resource_storage_->import_state_interfaces(resource_storage_->actuators_.back());
   resource_storage_->import_command_interfaces(
     resource_storage_->actuators_.back(), claimed_command_interface_map_);
@@ -271,9 +271,9 @@ size_t ResourceManager::actuator_components_size() const
   return resource_storage_->actuators_.size();
 }
 
-void ResourceManager::import_component(std::unique_ptr<components::SensorInterface> sensor)
+void ResourceManager::import_component(std::unique_ptr<SensorInterface> sensor)
 {
-  resource_storage_->sensors_.emplace_back(components::Sensor(std::move(sensor)));
+  resource_storage_->sensors_.emplace_back(Sensor(std::move(sensor)));
   resource_storage_->import_state_interfaces(resource_storage_->sensors_.back());
 }
 
@@ -282,9 +282,9 @@ size_t ResourceManager::sensor_components_size() const
   return resource_storage_->sensors_.size();
 }
 
-void ResourceManager::import_component(std::unique_ptr<components::SystemInterface> system)
+void ResourceManager::import_component(std::unique_ptr<SystemInterface> system)
 {
-  resource_storage_->systems_.emplace_back(components::System(std::move(system)));
+  resource_storage_->systems_.emplace_back(System(std::move(system)));
   resource_storage_->import_state_interfaces(resource_storage_->systems_.back());
   resource_storage_->import_command_interfaces(
     resource_storage_->systems_.back(), claimed_command_interface_map_);
