+++ hardware_interface/test/test_components/test_actuator.cpp
@@ -15,15 +15,15 @@
 #include <memory>
 #include <vector>
 
-#include "hardware_interface/components/base_interface.hpp"
-#include "hardware_interface/components/actuator_interface.hpp"
+#include "hardware_interface/base_interface.hpp"
+#include "hardware_interface/actuator_interface.hpp"
 
-using hardware_interface::status;
+using hardware_interface::ActuatorInterface;
+using hardware_interface::BaseInterface;
+using hardware_interface::CommandInterface;
 using hardware_interface::return_type;
 using hardware_interface::StateInterface;
-using hardware_interface::CommandInterface;
-using hardware_interface::components::ActuatorInterface;
-using hardware_interface::components::BaseInterface;
+using hardware_interface::status;
 
 class TestActuator : public BaseInterface<ActuatorInterface>
 {
@@ -110,4 +110,4 @@ private:
 };
 
 #include "pluginlib/class_list_macros.hpp"  // NOLINT
-PLUGINLIB_EXPORT_CLASS(TestActuator, hardware_interface::components::ActuatorInterface)
+PLUGINLIB_EXPORT_CLASS(TestActuator, hardware_interface::ActuatorInterface)
