+++ hardware_interface/test/test_components/test_components.xml
@@ -1,18 +1,18 @@
 <library path="test_components">
 
-  <class name="test_actuator" type="TestActuator" base_class_type="hardware_interface::components::ActuatorInterface">
+  <class name="test_actuator" type="TestActuator" base_class_type="hardware_interface::ActuatorInterface">
     <description>
       Test Actuator
     </description>
   </class>
 
-  <class name="test_sensor" type="TestSensor" base_class_type="hardware_interface::components::SensorInterface">
+  <class name="test_sensor" type="TestSensor" base_class_type="hardware_interface::SensorInterface">
     <description>
       Test Sensor
     </description>
   </class>
 
-  <class name="test_system" type="TestSystem" base_class_type="hardware_interface::components::SystemInterface">
+  <class name="test_system" type="TestSystem" base_class_type="hardware_interface::SystemInterface">
     <description>
       Test System
     </description>
