+++ hardware_interface/test/test_components/test_system.cpp
@@ -16,16 +16,16 @@
 #include <memory>
 #include <vector>
 
-#include "hardware_interface/components/base_interface.hpp"
-#include "hardware_interface/components/system_interface.hpp"
+#include "hardware_interface/base_interface.hpp"
+#include "hardware_interface/system_interface.hpp"
 #include "hardware_interface/types/hardware_interface_type_values.hpp"
 
-using hardware_interface::status;
-using hardware_interface::return_type;
+using hardware_interface::BaseInterface;
 using hardware_interface::CommandInterface;
+using hardware_interface::return_type;
 using hardware_interface::StateInterface;
-using hardware_interface::components::BaseInterface;
-using hardware_interface::components::SystemInterface;
+using hardware_interface::status;
+using hardware_interface::SystemInterface;
 
 class TestSystem : public BaseInterface<SystemInterface>
 {
@@ -90,4 +90,4 @@ private:
 };
 
 #include "pluginlib/class_list_macros.hpp"  // NOLINT
-PLUGINLIB_EXPORT_CLASS(TestSystem, hardware_interface::components::SystemInterface)
+PLUGINLIB_EXPORT_CLASS(TestSystem, hardware_interface::SystemInterface)
