+++ test_robot_hardware/src/test_force_torque_sensor.cpp
@@ -17,14 +17,14 @@
 #include <memory>
 #include <vector>
 
-#include "hardware_interface/components/base_interface.hpp"
-#include "hardware_interface/components/sensor_interface.hpp"
+#include "hardware_interface/base_interface.hpp"
+#include "hardware_interface/sensor_interface.hpp"
 
 using hardware_interface::status;
 using hardware_interface::return_type;
 using hardware_interface::StateInterface;
-using hardware_interface::components::BaseInterface;
-using hardware_interface::components::SensorInterface;
+using hardware_interface::BaseInterface;
+using hardware_interface::SensorInterface;
 
 namespace test_robot_hardware
 {
@@ -131,4 +131,4 @@ private:
 
 #include "pluginlib/class_list_macros.hpp"  // NOLINT
 PLUGINLIB_EXPORT_CLASS(
-  test_robot_hardware::TestForceTorqueSensor, hardware_interface::components::SensorInterface)
+  test_robot_hardware::TestForceTorqueSensor, hardware_interface::SensorInterface)
